refactor: migrate to mjs files
