+++ src/cli/args.mjs
@@ -1,0 +1,8 @@
+export const getValueByCLIArgs = (findName = '--') => { 
+    const fullArg = process.argv.find(el => el.indexOf(findName) === 0);
+    const idxOfEqual = fullArg.indexOf('=');
+    if (idxOfEqual >= 0) {
+        return fullArg.slice(idxOfEqual + 1)
+    } 
+    return undefined;
+};
