+++ src/cli/path.mjs
@@ -1,4 +1,4 @@
-import fs from 'node:fs/promises';
+import fs from 'node:fs/promises';  
 import path from 'node:path';
 
 export const isPathExist = async (pathForCheck) => {
@@ -32,4 +32,18 @@ export const getNewPathFromInput = async (inputPath, pathToCurrentDir, isNewPath
     }
 
     return newPath;
+}
+
+export const getUpPath = (pathToUp) => {
+    const arrOfDirNames = pathToUp.split(path.sep);
+    return arrOfDirNames.slice(0, -1).join(path.sep);  
+}
+export const getElTypeFromPath = async (pathToEl) => {
+    let isDir, isFile = false;
+    const stat = await fs.lstat(pathToEl);
+    isDir = stat.isDirectory();
+    if (isDir) return 'directory';
+    isFile = stat.isFile();
+    if (isFile) return 'file';
+    return undefined;
 }
