--- src/fs/delete.js
@@ -1,8 +1,0 @@
-import { unlink } from 'node:fs/promises';
-
-const deleteFile = async (pathToDeleteFile) => {
-    await unlink(pathToDeleteFile);
-    process.stdout.write(`${pathToDeleteFile} file deleted.\n`)
-};
-
-export default deleteFile;
