--- src/fs/files/dontLookAtMe.txt
@@ -1,1 +1,0 @@
-What are you looking at?!
