--- src/fs/files/fileToRead.txt
@@ -1,7 +1,0 @@
-My content
-should
-be
-printed
-into
-console
-!
