--- src/fs/files/fileToRemove.txt
@@ -1,1 +1,0 @@
-How dare you!
