--- src/fs/files/hello.txt
@@ -1,1 +1,0 @@
-Hello Node.js
