--- src/fs/files/wrongFilename.txt
@@ -1,3 +1,0 @@
-# This is a file with a wrong filename
-
-Hello from **markdown**!
