+++ src/hash/calcHash.mjs
