--- src/modules/files/b.json
@@ -1,5 +1,0 @@
-{
-    "a": 11,
-    "b": 22,
-    "c": 33
-}
