--- src/modules/files/c.js
@@ -1,1 +1,0 @@
-console.log('Hello from c.js!');
