--- src/streams/files/fileToRead.txt
@@ -1,1 +1,0 @@
-This file should be read using Streams API
